upgrade zktrie
