+++ Cargo.toml
@@ -8,7 +8,7 @@ log = "0.4"
 hex = "0.4"
 eth-types = { git = "https://github.com/scroll-tech/zkevm-circuits", features = ["scroll"], branch = "feat/rkyv" }
 mpt-zktrie = { git = "https://github.com/scroll-tech/zkevm-circuits", branch = "feat/rkyv" }
-zktrie = { git = "https://github.com/scroll-tech/zktrie.git", branch = "next", features= ["rs_zktrie"] }
+zktrie = { git = "https://github.com/scroll-tech/zktrie.git", branch = "main", features= ["rs_zktrie"] }
 rkyv = { version = "0.7", features = ["validation"] }
 itertools = "0.13.0" # for compatibility msrv 1.75
 
@@ -53,9 +53,6 @@ features = [
     "std",
     "optional_no_base_fee"
 ]
-
-[patch."https://github.com/scroll-tech/zktrie.git"]
-zktrie = { git = "https://github.com/scroll-tech//zktrie.git", branch = "next" }
 
 [[bin]]
 name = "stateless-block-verifier"
