+++ src/executor/mod.rs
@@ -7,7 +7,8 @@ use revm::{
     primitives::{AccountInfo, Env, SpecId},
 };
 use std::fmt::Debug;
-use zktrie::ZkTrie;
+use std::rc::Rc;
+use zktrie::{UpdateDb, ZkMemoryDb, ZkTrie};
 
 mod builder;
 /// Execute hooks
@@ -19,7 +20,8 @@ pub use builder::EvmExecutorBuilder;
 /// EVM executor that handles the block.
 pub struct EvmExecutor {
     db: CacheDB<ReadOnlyDB>,
-    zktrie: ZkTrie,
+    zktrie_db: Rc<ZkMemoryDb>,
+    zktrie: ZkTrie<UpdateDb>,
     spec_id: SpecId,
     hooks: hooks::ExecuteHooks,
 }
@@ -135,8 +137,7 @@ impl EvmExecutor {
                 // get storage tire
                 cycle_tracker_start!("update storage_tire");
                 let mut storage_tire = self
-                    .zktrie
-                    .get_db()
+                    .zktrie_db
                     .new_trie(storage_root_before.as_fixed_bytes())
                     .expect("unable to get storage trie");
                 for (key, value) in db_acc.storage.iter() {
